ET-465: fix commit hook for javascript files. It now runs through prettier to (#49)
fix formatting. Then it runs through eslint to check for errors. It
won't let you commit with eslint errors. It will also restage the file
that has prettier applied to it so nothing should fall through the
cracks. Also it is much faster because it only runs agains staged files.